requirements.txt
PySide6>=6.5.0
mysql-connector-python>=8.0.33

